% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/butterworth-filter.R
\name{compute_filter_args_butterworth_filter}
\alias{compute_filter_args_butterworth_filter}
\title{Compute filter arguments for the Butterworth filter based on parameters phi}
\usage{
compute_filter_args_butterworth_filter(phi, x)
}
\arguments{
\item{phi}{list of parameters including n, W}

\item{x}{vector to apply filtration to}
}
\value{
named list with parameters for a call to signal::filter.  Filter
  coefficients are computed via signal::remez
}
\description{
Compute filter arguments for the Butterworth filter based on parameters phi
}

