% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{compute_filter_values}
\alias{compute_filter_values}
\title{Compute filtered values}
\usage{
compute_filter_values(data, filter_control, phi)
}
\arguments{
\item{data}{a data frame with original data (no filtering done)}

\item{phi}{a list with one component for each component of the filter
function.  This component is a named list with arguments to the
corresponding filter function.}

\item{kcde_control}{a list of kcde_control parameters for kcde}

\item{log}{boolean; if TRUE (default), return filter values on the log scale}
}
\description{
Compute filtered values
}

