% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel-interaction.R
\name{compute_kernel_values}
\alias{compute_kernel_values}
\title{Compute kernel values measuring the similarity of each row in the
train_lagged_obs data frame to the prediction_lagged_obs.}
\usage{
compute_kernel_values(train_obs, prediction_obs, kernel_components, theta,
  log = TRUE)
}
\arguments{
\item{train_obs}{a data frame with lagged observation vectors computed
from the training data}

\item{prediction_obs}{a data frame with the lagged observation vector
computed from the prediction data.  It is assumed that
prediction_lagged_obs contains only one row.}

\item{theta}{a list with one component for each component of the kernel
function.  This component is a named list with arguments to the
corresponding kernel function.}

\item{log}{boolean; if TRUE (default), return kernel values on the log scale}

\item{kcde_control}{a list of kcde_control parameters for kcde}
}
\description{
Compute kernel values measuring the similarity of each row in the
train_lagged_obs data frame to the prediction_lagged_obs.
}

