% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{compute_na_rows_after_filter_and_offset}
\alias{compute_na_rows_after_filter_and_offset}
\title{Compute rows of the data frame that contain NA values after performing the
filtering and observation offset steps.  Also include values of
prediction_inds_not_included that were specified in kcde_control, if
applicable.}
\usage{
compute_na_rows_after_filter_and_offset(data, phi, vars_and_offsets,
  kcde_control)
}
\arguments{
\item{data}{data frame of observations to filter and offset}

\item{phi}{list of parameters for filtering}

\item{vars_and_offsets}{data frame describing variables to offset and size
and direction of offset}

\item{kcde_control}{list of control parameters for kcde}
}
\value{
integer vector with rows of data that contain NA values after
    filtering and offsetting observations
}
\description{
Compute rows of the data frame that contain NA values after performing the
filtering and observation offset steps.  Also include values of
prediction_inds_not_included that were specified in kcde_control, if
applicable.
}

