% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{compute_normalized_log_weights}
\alias{compute_normalized_log_weights}
\title{Normalize a vector of weights on log scale: given a vector of log_weights
such that exp(log_weights) are proportional to the final weights, update
so that exp(log_weights) sums to 1.}
\usage{
compute_normalized_log_weights(log_weights)
}
\arguments{
\item{log_weights:}{a vector of log(w) where w is proportional to the weights}
}
\value{
normalized log_weights so that sum(exp(log_weights)) = 1
}
\description{
Normalize a vector of weights on log scale: given a vector of log_weights
such that exp(log_weights) are proportional to the final weights, update
so that exp(log_weights) sums to 1.
}

