% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdtmvn-kernel-fn.R
\name{compute_pdtmvn_kernel_bw_params_from_bw_eigen}
\alias{compute_pdtmvn_kernel_bw_params_from_bw_eigen}
\title{Compute the parameters bw, bw_continuous, conditional_bw_discrete, and
conditional_center_discrete_offset_multiplier for the pdtmvn_kernel function
from the eigen-decomposition of the bandwidth matrix.}
\usage{
compute_pdtmvn_kernel_bw_params_from_bw_eigen(multilogit_bw_evecs, bw_evecs,
  log_bw_evals, bw_evals, continuous_var_col_inds, discrete_var_col_inds)
}
\arguments{
\item{bw_evecs}{matrix whose columns are the eigenvectors of the bandwidth
matrix.}

\item{bw_evals}{vector of eigenvalues of the bandwidth matrix.}

\item{continuous_vars}{Vector containing column indices for continuous
variables.}

\item{discrete_vars}{Vector containing column indices for discrete variables.}
}
\value{
Named list with four components: bw, bw_continuous,
    conditional_bw_discrete, and conditional_center_discrete_offset_multiplier
}
\description{
Compute the parameters bw, bw_continuous, conditional_bw_discrete, and
conditional_center_discrete_offset_multiplier for the pdtmvn_kernel function
from the eigen-decomposition of the bandwidth matrix.
}

