% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{create_kcde_control}
\alias{create_kcde_control}
\title{Assemble a list of kcde_control parameters for the kcde function with
    user-specified values.}
\usage{
create_kcde_control(X_names, y_names, time_name, prediction_horizons,
  kernel_components, filter_control, crossval_buffer,
  prediction_inds_not_included, loss_fn, loss_fn_prediction_args, loss_args,
  variable_selection_method = "stepwise", par_packages = NULL,
  par_cores = 1L, na.action = "na.omit")
}
\arguments{
\item{X_names}{a character vector of length >= 1 containing names of
variables in the data data frame to use in forming the lagged
observation process used for calculating weights}

\item{y_names}{a character vector of length 1 containing the name of the
variable in the data data frame to use as the target for prediction}

\item{prediction_horizons}{integer vector: the number of time steps between
the last observation and the time at which we make a prediction}

\item{kernel_components}{a list with one component for each component of the
kernel function.  Each component is a list with the following entries:
  - vars_and_lags: a data frame with two columns: "var_name" and
    "lag_value".  Each row specifies a combination of variable and lag
    that is included in this component of the kernel function.
  - kernel_fn: a function to evaluate the kernel
  - theta_fixed: a named list of parameters to kernel_fn whose values are
    held fixed (i.e., not estimated)
  - theta_est: a named list of parameters to kernel_fn whose values are
    to be estimated
  - initialize_theta_fn: a function to initialize both theta_fixed and
    theta_est
  - initialize_theta_args: a named list of arguments to
    initialize_theta_fn
  - vectorize_theta_est: a function that converts theta_est into an
    ordered vector on a scale suitable for passing as the first argument
    to optim.  Required to return a list with three components:
      (1) theta_est - vector of parameters to be estimated
      (2) lb - vector of lower bounds to theta_est
      (3) ub - vector of upper bounds to theta_est
  - update_theta_from_vectorized_theta_est: a function that updates
    theta_est (in list form) from theta_est (in vector form).}

\item{crossval_buffer}{during cross-validation, the number of indices before
the time at which we are making a prediction to drop from the "training
examples".}

\item{prediction_inds_not_included}{an integer vector specifying indices of
the data set that is provided in estimation that should not be included
in the calculation of the loss.  The observations at these indices may
still be included as conditioning variables used to make predictions at
other times.}

\item{loss_args}{a named list giving arguments to the loss function}

\item{variable_selection_method}{a character string specifying the method used to determine variable inclusion}

\item{par_packages}{a character vector containing names of packages that need
to be loaded in instances of R when computations are performed in
parallel.}

\item{na.action}{a character string specifying how NA values should be handled.}

\item{loss_fn_name}{a string giving the name of the function use to
compute loss from predictions}
}
\value{
the (at this point, unvalidated) list of kcde_control parameters
}
\description{
Assemble a list of kcde_control parameters for the kcde function with
    user-specified values.
}

