% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{create_kcde_control_default}
\alias{create_kcde_control_default}
\title{Assemble a list of kcde_control parameters for the kcde function with default
    values}
\usage{
create_kcde_control_default(X_names, y_names, data)
}
\arguments{
\item{X_names}{a character vector of length >= 1 containing names of
variables in the data data frame to use in forming the lagged
observation process used for calculating weights}

\item{y_names}{a character vector of length 1 containing the name of the
variable in the data data frame to use as the target for prediction}

\item{data}{a data frame where rows are consecutive observations}
}
\value{
the list of kcde_control parameters
}
\description{
Assemble a list of kcde_control parameters for the kcde function with default
    values
}

