% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{est_kcde_params_stepwise_crossval}
\alias{est_kcde_params_stepwise_crossval}
\title{Use a stepwise procedure to estimate parameters and select lags by optimizing
a cross-validation estimate of predictive performance}
\usage{
est_kcde_params_stepwise_crossval(data, kcde_control, init_theta_vector,
  init_phi_vector)
}
\arguments{
\item{data}{the data frame to use in performing cross validation}

\item{kcde_control}{a list of parameters specifying how the fitting is done}
}
\value{
a list with two components: vars_and_offsets is the estimated "optimal" lags
    to use for each variable, and theta_hat is the estimated "optimal"
    kernel parameters to use for each combination of variable and lag
}
\description{
Use a stepwise procedure to estimate parameters and select lags by optimizing
a cross-validation estimate of predictive performance
}

