% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{est_kcde_params_stepwise_crossval_one_potential_step}
\alias{est_kcde_params_stepwise_crossval_one_potential_step}
\title{Estimate the parameters theta and corresponding cross-validated estimate of
loss for one possible model obtained by adding or removing a variable/lag
combination from the model obtained in the previous iteration of the stepwise
search procedure.}
\usage{
est_kcde_params_stepwise_crossval_one_potential_step(prev_vars_and_offsets,
  prev_theta, prev_phi, init_theta_vector, init_phi_vector, update_var_name,
  update_lag_value, data, all_na_drop_rows, kcde_control)
}
\arguments{
\item{prev_vars_and_offsets}{list representing combinations of variables and lags
included in the model obtained at the previous step}

\item{prev_theta}{list representing the kernel parameter estimates obtained
at the previous step}

\item{update_var_name}{the name of the variable to try adding or removing
from the model}

\item{data}{the data frame with observations used in estimating model
parameters}

\item{kcde_control}{a list of parameters specifying how the fitting is done}

\item{update_offset_value}{the value of the offset for the variable specified by
update_var_name to try adding or removing from the model}
}
\value{
a list with three components: loss is a cross-validation estimate of
    the loss associated with the estimated parameter values for the given
    model, lags is a list representing combinations of variables and lags
    included in the updated model, and theta is a list representing the
    kernel parameter estimates in the updated model
}
\description{
Estimate the parameters theta and corresponding cross-validated estimate of
loss for one possible model obtained by adding or removing a variable/lag
combination from the model obtained in the previous iteration of the stepwise
search procedure.
}

