% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{extract_vectorized_phi_est_from_phi}
\alias{extract_vectorized_phi_est_from_phi}
\title{Extract a vector of parameter values that are to be estimated from phi,
on the estimation scale.}
\usage{
extract_vectorized_phi_est_from_phi(phi, filter_control)
}
\arguments{
\item{phi}{filtering parameters phi in list form}

\item{filter_control}{control parameters for the kcde fit}
}
\value{
numeric vector with parameter values
}
\description{
Extract a vector of parameter values that are to be estimated from phi,
on the estimation scale.
}

