% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel-interaction.R
\name{extract_vectorized_theta_est_from_theta}
\alias{extract_vectorized_theta_est_from_theta}
\title{Extract a vector of parameter values that are to be estimated from theta,
on the estimation scale.}
\usage{
extract_vectorized_theta_est_from_theta(theta, vars_and_offsets, kcde_control)
}
\arguments{
\item{kcde_control}{control parameters for the kcde fit}

\item{theta_list}{kernel parameters theta in list form}

\item{lags}{list representing combinations of variables and lags
included in the model}
}
\value{
numeric vector with parameter values
}
\description{
Extract a vector of parameter values that are to be estimated from theta,
on the estimation scale.
}

