% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdtmvn-kernel-fn.R
\name{get_col_inds_continuous_discrete_vars_used}
\alias{get_col_inds_continuous_discrete_vars_used}
\title{Create two integer vectors with indices of columns in x corresponding to
continuous variables and discrete variables.  These integer vectors are
suitable for passing to functions from the pdtmvn function.}
\usage{
get_col_inds_continuous_discrete_vars_used(x_colnames, continuous_vars,
  discrete_vars)
}
\arguments{
\item{x_colnames}{character vector with column names of x}

\item{continuous_vars}{character vector with names of columns in x that
should be treated as continuous variables}

\item{discrete_vars}{character vector with names of columns in x that
should be treated as discrete variables}
}
\value{
list with two components: continuous_vars is an integer vector of
    columns in x corresponding to continuous variables and discrete_vars is
    an integer vector of columns in x corresponding to discrete variables
}
\description{
Create two integer vectors with indices of columns in x corresponding to
continuous variables and discrete variables.  These integer vectors are
suitable for passing to functions from the pdtmvn function.
}

