% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{get_default_kernel_components}
\alias{get_default_kernel_components}
\title{Get default kernel functions based on a brief look at the data.  This is
unreliable.  Update to return periodic_kernel if X_names[i] == time_name?}
\usage{
get_default_kernel_components(X_names, y_names, data)
}
\arguments{
\item{X_names}{a character vector of length >= 1 containing names of
variables in the data data frame to use in forming the lagged
observation process used for calculating weights}

\item{y_names}{a character vector of length 1 containing the name of the
variable in the data data frame to use as the target for prediction}

\item{data}{a data frame where rows are consecutive observations}

\item{time_name}{(optional) a character vector of length 1 containing the
name of the variable in the data data frame to use as the time.}
}
\value{
a list of default parameters for kernel components -- probably all bad
}
\description{
Get default kernel functions based on a brief look at the data.  This is
unreliable.  Update to return periodic_kernel if X_names[i] == time_name?
}

