% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_inds_smallest_k}
\alias{get_inds_smallest_k}
\title{Get the indices of the smallest k elements of v.}
\usage{
get_inds_smallest_k(v, k)
}
\arguments{
\item{v}{a vector}

\item{k}{number of indices to return}
}
\value{
a vector of length k containing the indices of the k smallest
    elements of v, in ascending order.
}
\description{
Get the indices of the smallest k elements of v.
}

