% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{get_phi_optim_bounds}
\alias{get_phi_optim_bounds}
\title{Get vectors of lower and upper bounds for phi parameters to be used in
calls to optim.}
\usage{
get_phi_optim_bounds(phi, filter_control)
}
\arguments{
\item{phi}{list of filter parameters phi that are being estimated}

\item{filter_control}{control parameters for filtering}
}
\value{
list with two components: lower and upper, giving vectors
  with lower and upper bounds for possible parameter values
}
\description{
Get vectors of lower and upper bounds for phi parameters to be used in
calls to optim.
}

