% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdtmvn-kernel-fn.R
\name{get_theta_optim_bounds_pdtmvn_kernel}
\alias{get_theta_optim_bounds_pdtmvn_kernel}
\title{Get lower and upper bounds for the theta parameters being estimated
in the pdtmvn_kernel}
\usage{
get_theta_optim_bounds_pdtmvn_kernel(theta_list, ...)
}
\arguments{
\item{theta_list}{parameters to pdtmvn kernel in list format}

\item{...}{mop up arguments}
}
\value{
list with two components: lower and upper, numeric vectors
}
\description{
Get lower and upper bounds for the theta parameters being estimated
in the pdtmvn_kernel
}

