% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodic-kernel-fn.R
\name{get_theta_optim_bounds_periodic_kernel}
\alias{get_theta_optim_bounds_periodic_kernel}
\title{Get lower and upper bounds for the log_bw parameter in the periodic_kernel}
\usage{
get_theta_optim_bounds_periodic_kernel(...)
}
\arguments{
\item{...}{mop up arguments}
}
\value{
list with two components: lower and upper, numeric vectors
}
\description{
Get lower and upper bounds for the log_bw parameter in the periodic_kernel
}

