% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/butterworth-filter.R
\name{initialize_filter_params_butterworth_filter}
\alias{initialize_filter_params_butterworth_filter}
\title{Initialize parameters of the pm_opt_filter on evaluation scale.
In principle, we could/should do something data based here?  Instead, I just made up
default numbers.}
\usage{
initialize_filter_params_butterworth_filter(prev_phi, x, ...)
}
\arguments{
\item{prev_phi}{list of previous filtering parameters for the given random variable}

\item{x}{vector of observations for the given random variable to be filtered}

\item{...}{absorb arguments}
}
\value{
initialized list of parameters
}
\description{
Initialize parameters of the pm_opt_filter on evaluation scale.
In principle, we could/should do something data based here?  Instead, I just made up
default numbers.
}

