% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodic-kernel-fn.R
\name{initialize_params_periodic_kernel}
\alias{initialize_params_periodic_kernel}
\title{Get initial parameter values for the periodic_kernel}
\usage{
initialize_params_periodic_kernel(x, prev_theta, total_num_vars, sample_size,
  ...)
}
\arguments{
\item{x}{a matrix of values used to initialize the parameter values for the
kernel function.  Each row is an observation, each column is an
observed variable.}

\item{...}{used to absorb other arguments in the function call}

\item{continuous_vars}{character vector with variables names that are to be
treated as continuous}

\item{discrete_vars}{character vector with variables names that are to be
treated as discrete}

\item{kcde_control}{list of control parameters to kcde}
}
\value{
list with initial values of parameters to the periodic_kernel
}
\description{
Get initial parameter values for the periodic_kernel
}

