% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{initialize_phi}
\alias{initialize_phi}
\title{Initialize parameter values for filtering}
\usage{
initialize_phi(prev_phi, updated_vars_and_offsets, update_var_name,
  update_offset_value, update_offset_type, data, kcde_control)
}
\arguments{
\item{prev_phi}{previous phi values before update to variables and lags}

\item{updated_vars_and_offsets}{lags after update}

\item{update_var_name}{character; the name of the variable added/removed from the model}

\item{update_offset_value}{integer; the offset that was added/removed from the model}

\item{update_offset_type}{integer; the type of the offset that was added/removed from the model}

\item{data}{data matrix}

\item{kcde_control}{list of control parameters for kcde}
}
\value{
list of phi parameters
}
\description{
Initialize parameter values for filtering
}

