% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{kcde}
\alias{kcde}
\title{Estimate the parameters for kcde.  There is redundancy here in that X_names,
y_names, time_name are all included in the kcde_control object as well as
parameters to this function.  Decide on an interface.}
\usage{
kcde(X_names, y_names, data, init_theta_vector, init_phi_vector, kcde_control)
}
\arguments{
\item{X_names}{a character vector of length >= 1 containing names of
variables in the data data frame to use in forming the lagged
observation process used for calculating weights}

\item{y_names}{a character vector of length 1 containing the name of the
variable in the data data frame to use as the target for prediction}

\item{data}{a data frame where rows are consecutive observations}

\item{kcde_control}{a list of parameters kcde_controlling how the fit is done.
See the documentation for kcde_control.}
}
\value{
an object representing an estimated kcde model.  Currently, a list
    with 7 components.
}
\description{
Estimate the parameters for kcde.  There is redundancy here in that X_names,
y_names, time_name are all included in the kcde_control object as well as
parameters to this function.  Decide on an interface.
}

