% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{kcde_predict}
\alias{kcde_predict}
\title{Make predictions from an estimated kcde model forward prediction_horizon time
steps from the end of predict_data, based on the weighting variables, lags,
kernel functions, and bandwidths specified in the kcde_fit object.}
\usage{
kcde_predict(kcde_fit, prediction_data,
  leading_rows_to_drop = max(kcde_fit$vars_and_offsets$offset_value[kcde_fit$vars_and_offsets$offset_type
  == "lag"]),
  trailing_rows_to_drop = max(kcde_fit$vars_and_offsets$offset_value[kcde_fit$vars_and_offsets$offset_type
  == "horizon"]), additional_training_rows_to_drop = NULL,
  prediction_type = "distribution", n, p, q, prediction_test_lead_obs,
  log = FALSE)
}
\arguments{
\item{kcde_fit}{is an object representing a fitted kcde model}

\item{prediction_data}{is a vector of data points to use in prediction}

\item{prediction_type}{character; either "distribution" or "point",
indicating the type of prediction to perform.}

\item{normalize_weights}{boolean, should the weights be normalized?}
}
\value{
an object with prediction results; the contents depend on the value
    of prediction_type
}
\description{
Make predictions from an estimated kcde model forward prediction_horizon time
steps from the end of predict_data, based on the weighting variables, lags,
kernel functions, and bandwidths specified in the kcde_fit object.
}

