% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{kcde_predict_given_lagged_obs}
\alias{kcde_predict_given_lagged_obs}
\title{Make predictions from an estimated kcde model forward prediction_horizon time
steps from the end of predict_data, based on the kernel functions and
bandwidths specified in the kcde_fit object.  This function requires that the
lagged and lead observation vectors have already been computed.}
\usage{
kcde_predict_given_lagged_obs(train_lagged_obs, train_lead_obs,
  prediction_lagged_obs, prediction_test_lead_obs, kcde_fit,
  prediction_type = "distribution", n, p, q, log = FALSE)
}
\arguments{
\item{train_lagged_obs}{is a matrix (with column names) containing the
lagged observation vector computed from the training data.  Each row
corresponds to a time point.  Each column is a (lagged) variable.}

\item{train_lead_obs}{is a vector with length = nrow(train_lagged_obs) with
the value of the prediction target variable corresponding to each row in
the train_lagged_obs matrix.}

\item{prediction_lagged_obs}{is a matrix (with column names) containing the
lagged observation vector computed from the prediction data.  There is
only one row, representing one time point.  Each column is a (lagged)
variable.}

\item{kcde_fit}{is an object representing a fitted kcde model}

\item{prediction_type}{character; either "distribution" or "point",
indicating the type of prediction to perform.}
}
\value{
an object with prediction results; the contents depend on the value
    of prediction_type
}
\description{
Make predictions from an estimated kcde model forward prediction_horizon time
steps from the end of predict_data, based on the kernel functions and
bandwidths specified in the kcde_fit object.  This function requires that the
lagged and lead observation vectors have already been computed.
}

