% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{kcde_prob_predict_given_lagged_lead_obs}
\alias{kcde_prob_predict_given_lagged_lead_obs}
\title{Compute the conditional (predictive) probability that the predictive target variables
are all less than or equal to the corresponding elements of the rows of q.}
\usage{
kcde_prob_predict_given_lagged_lead_obs(q, n, train_lagged_obs, train_lead_obs,
  prediction_lagged_obs, kcde_fit)
}
\arguments{
\item{n}{sample size for sample from predictive distribution used in
approximating quantiles}

\item{train_lagged_obs}{is a matrix (with column names) containing the
lagged observation vector computed from the training data.  Each row
corresponds to a time point.  Each column is a (lagged) variable.}

\item{train_lead_obs}{is a vector with length = nrow(train_lagged_obs) with
the value of the prediction target variable corresponding to each row in
the train_lagged_obs matrix.}

\item{prediction_lagged_obs}{is a matrix (with column names) containing the
lagged observation vector computed from the prediction data.  There is
only one row, representing one time point.  Each column is a (lagged)
variable.}

\item{kcde_fit}{is an object representing a fitted kcde model}

\item{p}{vector of probabilities which to compute quantiles}
}
\value{
a vector of probabilities
}
\description{
Compute the conditional (predictive) probability that the predictive target variables
are all less than or equal to the corresponding elements of the rows of q.
}

