% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{kcde_sample_predict_given_lagged_lead_obs}
\alias{kcde_sample_predict_given_lagged_lead_obs}
\title{Draw a sample from the predictive distribution corresponding to an estimated
kcde model forward prediction_horizon time steps from the end of predict_data.
This function requires that the kernel weights and centers have already been
computed.}
\usage{
kcde_sample_predict_given_lagged_lead_obs(n, train_lagged_obs, train_lead_obs,
  prediction_lagged_obs, kcde_fit)
}
\arguments{
\item{n}{sample size for sample from predictive distribution used in
approximating quantiles}

\item{train_lagged_obs}{is a matrix (with column names) containing the
lagged observation vector computed from the training data.  Each row
corresponds to a time point.  Each column is a (lagged) variable.}

\item{train_lead_obs}{is a vector with length = nrow(train_lagged_obs) with
the value of the prediction target variable corresponding to each row in
the train_lagged_obs matrix.}

\item{prediction_lagged_obs}{is a matrix (with column names) containing the
lagged observation vector computed from the prediction data.  There is
only one row, representing one time point.  Each column is a (lagged)
variable.}

\item{kcde_fit}{is an object representing a fitted kcde model}
}
\value{
a matrix with samples from the predictive distribution
}
\description{
Draw a sample from the predictive distribution corresponding to an estimated
kcde model forward prediction_horizon time steps from the end of predict_data.
This function requires that the kernel weights and centers have already been
computed.
}

