% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{log_score_loss}
\alias{log_score_loss}
\title{Compute log score based on a vector of log(p(data)) where p is a predictive distribution}
\usage{
log_score_loss(prediction_result, ...)
}
\description{
Compute log score based on a vector of log(p(data)) where p is a predictive distribution
}

