% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{logspace_add}
\alias{logspace_add}
\title{Add two values in logspace}
\usage{
logspace_add(logx, logy)
}
\arguments{
\item{logx}{first number}

\item{logy}{second number}
}
\value{
sum of the numbers
}
\description{
Add two values in logspace
}

