% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{logspace_sub}
\alias{logspace_sub}
\title{Subtract two values in logspace}
\usage{
logspace_sub(logx, logy)
}
\arguments{
\item{logx}{minuend}

\item{logy}{subtrahend}
}
\value{
result of the subtraction
}
\description{
Subtract two values in logspace
}

