% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{logspace_sub_matrix_rows}
\alias{logspace_sub_matrix_rows}
\title{Row-wise difference of matrix with two columns in logspace}
\usage{
logspace_sub_matrix_rows(logX)
}
\arguments{
\item{logX}{the matrix}
}
\value{
a vector with each element being the difference of corresponding
    elements in the first and second column
}
\description{
Row-wise difference of matrix with two columns in logspace
}

