% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{logspace_sum}
\alias{logspace_sum}
\title{Find sum of numeric vector in logspace}
\usage{
logspace_sum(logx)
}
\arguments{
\item{logx}{the vector to sum}
}
\value{
sum of all the elements in the vector
}
\description{
Find sum of numeric vector in logspace
}

