% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{logspace_sum_matrix_rows}
\alias{logspace_sum_matrix_rows}
\title{Row-wise sum of the given matrix in logspace}
\usage{
logspace_sum_matrix_rows(logX)
}
\arguments{
\item{logX}{the matrix to sum}
}
\value{
a vector with each element summing corresponding row of matrix
}
\description{
Row-wise sum of the given matrix in logspace
}

