% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{mae_from_kernel_weights_and_centers}
\alias{mae_from_kernel_weights_and_centers}
\title{Compute "mean" absolute error for one time point from prediction in the form
of kernel weights and centers.}
\usage{
mae_from_kernel_weights_and_centers(kernel_weights_and_centers, obs)
}
\arguments{
\item{kernel_weights_and_centers}{a named list with two components: weights
is a vector of kernel weights, and centers is a vector of kernel centers}

\item{obs}{is a numeric with length 1 containign the observed value for one
time point}
}
\value{
abs(obs - prediction) where prediction is the weighted mean of the
    kernel centers.
}
\description{
Compute "mean" absolute error for one time point from prediction in the form
of kernel weights and centers.
}

