% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{mase}
\alias{mase}
\title{Compute Mean Absolute Scaled Error (MASE) to evaluate point predictions}
\usage{
mase(obs, pred)
}
\arguments{
\item{obs}{vector of observed values}

\item{pred}{vector of point predictions}
}
\value{
mean absolute scaled error for given vectors
}
\description{
Compute Mean Absolute Scaled Error (MASE) to evaluate point predictions
}

