% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{mase_from_kernel_weights_and_centers}
\alias{mase_from_kernel_weights_and_centers}
\title{Compute MASE from predictions that are in the form of kernel weights and
centers.  This is currently broken because we need predictions for a whole
time series to compute MASE, and the current "implementation" only takes
values for one time point....}
\usage{
mase_from_kernel_weights_and_centers(kernel_weights_and_centers, obs)
}
\description{
Compute MASE from predictions that are in the form of kernel weights and
centers.  This is currently broken because we need predictions for a whole
time series to compute MASE, and the current "implementation" only takes
values for one time point....
}

