% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{neg_log_score_loss}
\alias{neg_log_score_loss}
\title{Compute negative log score based on a vector of log(p(data)) where p is a
predictive distribution}
\usage{
neg_log_score_loss(prediction_result, ...)
}
\description{
Compute negative log score based on a vector of log(p(data)) where p is a
predictive distribution
}

