% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{one_pass_signal_filter}
\alias{one_pass_signal_filter}
\title{One-pass filter using signal::filter, with two additions:
  1) We impute internal missing values for the purposes of filtering, then
     replace them with NAs again after filtering
  2) We pad the end of the time series with the last observed value instead
     of zeros.  Based on very informal plotting, this seems to give better
     values at the end of the time series, which will be important for
     prediction}
\usage{
one_pass_signal_filter(filt, x, impute_fn)
}
\arguments{
\item{x}{vector of data to be filtered}

\item{filter}{vector of filter coefficients for a FIR filter}

\item{method}{method for a call to stats::filter -- see the documentation at stats::filter}

\item{sides}{sides for a call to stats::filter -- see the documentation at stats::filter}

\item{circular}{circular for a call to stats::filter -- see the documentation at stats::filter}
}
\description{
One-pass filter using signal::filter, with two additions:
  1) We impute internal missing values for the purposes of filtering, then
     replace them with NAs again after filtering
  2) We pad the end of the time series with the last observed value instead
     of zeros.  Based on very informal plotting, this seems to give better
     values at the end of the time series, which will be important for
     prediction
}

