% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodic-kernel-fn.R
\name{periodic_kernel}
\alias{periodic_kernel}
\title{Evaluate the periodic kernel function}
\usage{
periodic_kernel(x, center, period, bw, log, ...)
}
\arguments{
\item{x}{a vector of values at which to evaluate the kernel function}

\item{center}{a real number, center point for the kernel function}

\item{period}{kernel period}

\item{bw}{kernel bandwidth}
}
\value{
vector of the kernel function value at each point in x
}
\description{
Evaluate the periodic kernel function
}

