% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-pdtmvn-mode-centered-kernel-fn.R
\name{rlog_pdtmvn_mode_centered_kernel}
\alias{rlog_pdtmvn_mode_centered_kernel}
\title{Simulate from the kernel function given by the pdtmvn distribution.}
\usage{
rlog_pdtmvn_mode_centered_kernel(n, conditioning_obs, center, bw, bw_continuous,
  conditional_bw_discrete, conditional_center_discrete_offset_multiplier,
  continuous_vars, discrete_vars, continuous_var_col_inds,
  discrete_var_col_inds, discrete_var_range_fns, lower, upper, x_names, ...)
}
\arguments{
\item{n}{number of simulations to generate}

\item{center}{a real vector, center point for the kernel function}

\item{bw}{bandwidth matrix}

\item{bw_continuous}{the portion of bw corresponding to continuous variables}

\item{conditional_bw_discrete}{the Schur complement of the portion of bw
corresponding to discrete variables}

\item{conditional_center_discrete_offset_multiplier}{Sigma_dc Sigma_c^{-1}.
This is used in computing the mean of the underlying multivariate normal
distribution for the discrete variables conditioning on the continuous
variables.}

\item{continuous_vars}{character vector with names of variables that are to
be treated as continuous.  May contain entries that do not appear in
colnames(x).}

\item{discrete_vars}{character vector with names of variables that are to be
treated as discrete.  May contain entries that do not appear in
colnames(x)}

\item{discrete_var_range_fns}{a list with one entry for each element
of discrete_vars.  Each entry is a named list of length 2; the element
named "a" is a character string with the name of a function that returns
a(x) for any real x, and the element named "b" is a character string with
the name of a function that returns b(x) for any real x.}

\item{lower}{Vector of lower truncation points}

\item{upper}{Vector of upper truncation points}

\item{...}{mop up extra arguments}

\item{a}{matrix of values at which to evaluate the kernel function, with
column names specified.  Each row is an observation, each column is an
observed variable.}

\item{log}{logical; if TRUE, return the log of the kernel function value}
}
\value{
the value of the kernel function given by the pdtmvn distribution
    at x.
}
\description{
Simulate from the kernel function given by the pdtmvn distribution.
}

