% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel-interaction.R
\name{simulate_values_from_product_kernel}
\alias{simulate_values_from_product_kernel}
\title{Simulate values from (conditional) kernel.}
\usage{
simulate_values_from_product_kernel(n, conditioning_obs, center,
  kernel_components, theta)
}
\arguments{
\item{theta}{a list with one component for each component of the kernel
function.  This component is a named list with arguments to the
corresponding kernel function.}

\item{train_obs}{a data frame with lagged observation vectors computed
from the training data}

\item{prediction_obs}{a data frame with the lagged observation vector
computed from the prediction data.  It is assumed that
prediction_lagged_obs contains only one row.}

\item{kcde_control}{a list of kcde_control parameters for kcde}

\item{log}{boolean; if TRUE (default), return kernel values on the log scale}
}
\description{
Simulate values from (conditional) kernel.
}

