% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{two_pass_signal_filter}
\alias{two_pass_signal_filter}
\title{Forward and reverse pass filter to correct phase shift introduced by one-pass
filter.  Uses signal::filter to do each filtering pass.  Closely based on
signal::filtfilt -- see the documentation there for further discussion.  Two
main differences with that implementation:
  1) We impute internal missing values for the purposes of filtering, then
     replace them with NAs again after filtering
  2) We pad the end of the time series with the last observed value instead
     of zeros.  Based on very informal plotting, this seems to give better
     values at the end of the time series, which will be important for
     prediction}
\usage{
two_pass_signal_filter(filt, x, impute_fn)
}
\arguments{
\item{x}{vector of data to be filtered}

\item{filter}{vector of filter coefficients for a FIR filter}

\item{method}{method for a call to stats::filter -- see the documentation at stats::filter}

\item{sides}{sides for a call to stats::filter -- see the documentation at stats::filter}

\item{circular}{circular for a call to stats::filter -- see the documentation at stats::filter}
}
\description{
Forward and reverse pass filter to correct phase shift introduced by one-pass
filter.  Uses signal::filter to do each filtering pass.  Closely based on
signal::filtfilt -- see the documentation there for further discussion.  Two
main differences with that implementation:
  1) We impute internal missing values for the purposes of filtering, then
     replace them with NAs again after filtering
  2) We pad the end of the time series with the last observed value instead
     of zeros.  Based on very informal plotting, this seems to give better
     values at the end of the time series, which will be important for
     prediction
}

