% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/butterworth-filter.R
\name{update_filter_params_from_vectorized_butterworth_filter}
\alias{update_filter_params_from_vectorized_butterworth_filter}
\title{Update the parameter list phi from a vector of parameters on estimation scale}
\usage{
update_filter_params_from_vectorized_butterworth_filter(phi_est_vector, phi)
}
\arguments{
\item{phi_est_vector}{a numeric vector of length at least 2}

\item{phi}{a list of parameters to update}
}
\value{
updated list phi
}
\description{
Update the parameter list phi from a vector of parameters on estimation scale
}

