% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{update_phi_from_vectorized_phi_est}
\alias{update_phi_from_vectorized_phi_est}
\title{Convert phi from vector form to list form}
\usage{
update_phi_from_vectorized_phi_est(phi_est_vector, phi, filter_control)
}
\arguments{
\item{phi_est_vector}{numeric vector of filter parameters phi that are
being estimated, on estimation scale.}

\item{kcde_control}{control parameters for the kcde fit}
}
\value{
list of lists of parameter values -- outer list has one component
    for each filter function, inner list has one component
    for each parameter used in the corresponding filter function
}
\description{
Convert phi from vector form to list form
}

