% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel-interaction.R
\name{update_theta_from_vectorized_theta_est}
\alias{update_theta_from_vectorized_theta_est}
\title{Convert theta from vector form to list form}
\usage{
update_theta_from_vectorized_theta_est(theta_est_vector, theta, kcde_control)
}
\arguments{
\item{theta_est_vector}{numeric vector of kernel parameters theta that are
being estimated, on estimation scale.}

\item{kcde_control}{control parameters for the kcde fit}
}
\value{
list of lists of parameter values -- outer list has one component
    for each kernel function, inner list has one component
    for each parameter used in the corresponding kernel function
}
\description{
Convert theta from vector form to list form
}

