% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodic-kernel-fn.R
\name{update_theta_from_vectorized_theta_est_periodic_kernel}
\alias{update_theta_from_vectorized_theta_est_periodic_kernel}
\title{A function to unvectorize the parameters of the pdtmvn_kernel and convert
from estimation scale to scale actually used.}
\usage{
update_theta_from_vectorized_theta_est_periodic_kernel(theta_est_vector, theta)
}
\arguments{
\item{theta_vector}{a numeric vector of parameters that are being optimized,
on a scale suitable for use in optim.}

\item{parameterization}{character; currently, only supported value is
"bw-diagonalized-est-eigenvalues"}

\item{bw_evecs}{square matrix with eigenvectors of the bandwidth matrix in
columns}

\item{continuous_vars}{character vector with variables names that are to be
treated as continuous}

\item{discrete_vars}{character vector with variables names that are to be
treated as discrete}

\item{kcde_control}{list of control parameters to kcde}
}
\value{
list of parameters to pdtmvn_kernel
}
\description{
A function to unvectorize the parameters of the pdtmvn_kernel and convert
from estimation scale to scale actually used.
}

