% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{update_vars_and_offsets}
\alias{update_vars_and_offsets}
\title{Update the list that keeps track of which combinations of variables and
offsets are included in the model by adding or removing (as appropriate) the
combination specified by update_var_name, update_offset_values, and
update_offset_type}
\usage{
update_vars_and_offsets(prev_vars_and_offsets, update_var_name,
  update_offset_value, update_offset_type)
}
\arguments{
\item{prev_vars_and_offsets}{list of previous variable/lag combinations to
update}

\item{update_var_name}{name of the variable to update}

\item{update_offset_value}{value of the offset to update}

\item{update_offset_type}{type of the offset to update}
}
\value{
updated list of variable/lag combinations
}
\description{
Update the list that keeps track of which combinations of variables and
offsets are included in the model by adding or removing (as appropriate) the
combination specified by update_var_name, update_offset_values, and
update_offset_type
}

