% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{validate_kcde_control}
\alias{validate_kcde_control}
\title{Validate kcde_control parameters for kcde -- not implemented}
\usage{
validate_kcde_control(kcde_control, X_names, y_names, time_name, data)
}
\arguments{
\item{kcde_control}{a list of kcde_control parameters for kcde}

\item{X_names}{a character vector of length >= 1 containing names of
variables in the data data frame to use in forming the lagged
observation process used for calculating weights}

\item{y_names}{a character vector of length 1 containing the name of the
variable in the data data frame to use as the target for prediction}

\item{time_name}{(optional) a character vector of length 1 containing the
name of the variable in the data data frame to use as the time.}

\item{data}{a data frame where rows are consecutive observations}
}
\value{
no return value -- either stops with an error or not.
}
\description{
Validate kcde_control parameters for kcde -- not implemented
}

