% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/butterworth-filter.R
\name{vectorize_filter_params_butterworth_filter}
\alias{vectorize_filter_params_butterworth_filter}
\title{Vectorize the parameters of the pm_opt_filter and convert
to estimation scale.}
\usage{
vectorize_filter_params_butterworth_filter(phi)
}
\arguments{
\item{phi}{list of parameters on evaluation scale}
}
\value{
vector of parameters on estimation scale
}
\description{
Vectorize the parameters of the pm_opt_filter and convert
to estimation scale.
}

