% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodic-kernel-fn.R
\name{vectorize_params_periodic_kernel}
\alias{vectorize_params_periodic_kernel}
\title{A function to vectorize the parameters of the pdtmvn_kernel and convert
to estimation scale.}
\usage{
vectorize_params_periodic_kernel(theta_list, parameterization, ...)
}
\arguments{
\item{theta_list}{parameters for the pdtmvn_kernel in list format}

\item{parameterization}{character; currently, only supported value is
"bw-diagonalized-est-eigenvalues"}

\item{...}{mop up other arguments}
}
\value{
vector containing parameters that are estimated on a scale
    suitable for numerical optimization
}
\description{
A function to vectorize the parameters of the pdtmvn_kernel and convert
to estimation scale.
}

