% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{mae}
\alias{mae}
\title{Compute Mean Absolute Error to evaluate point predictions}
\usage{
mae(obs, pred)
}
\arguments{
\item{obs}{vector of observed values}

\item{pred}{vector of point predictions}
}
\description{
Compute Mean Absolute Error to evaluate point predictions
}

